### 2020-01-07 b45e247
Update paper.md
