### 2020-01-08 6ac4256
Update paper.md
