### 2020-01-09 #1998 @arokem
Hmm. Looks like I broke things in my PR (https://github.com/pySRURGS/pyGOURGS/pull/1). But it wasn't working before either (the title was mis-rendered and the bibliography wasn't rendering)
