### 2020-01-09 cfe01ed
Update paper.md
