+++ papers/JOSS/paper.md
@@ -51,7 +51,7 @@ of random search can require differing amounts of computational effort to
 reach a high quality solution. They found that the random search method commonly 
 used to generate the initial population of genetic programming solutions performs 
 much worse than does uniform random search. We found one prominent modern paper 
-which claimed that genetic programming outperformed random search [@Sipper:2019],
+which claimed that genetic programming outperformed random search [@Sipper:2018],
 when in fact they were comparing genetic programming to a biased type of random 
 search which they then put on further unequal footing. This software has broad 
 applicability in the examination of the solution space for global optimization 
