### 2020-01-10 #1998 @arokem
👋 @andim: would you be willing to review this paper for JOSS?
