### 2020-01-10 #1998 @arokem
👋 @rowanc1 : would you be willing to review this paper for JOSS?
