### 2020-01-19 4fdd2d6
plotting all the trees for the ant problem for clarity
