### 2020-01-19 3cacfcc
printing plots
