### 2020-01-21 #1998 @arokem
@fernandoandreotti : would you be willing to review this paper for JOSS?
