### 2020-01-21 b5e83cd
script produces publication ready figur
