### 2020-01-21 b8e0f05
working on script to read output database
