### 2020-02-03 #1998 @arokem
@iljah : would you be willing to review this article for JOSS?
