### 2020-02-05 #2074 @whedon
[REVIEW]: pyGOURGS - global optimization of n-ary tree representable problems using uniform random global search
**Submitting author:** @pySRURGS (<a href="http://orcid.org/0000-0002-3050-8943">Sohrab Towfighi</a>)
**Repository:** <a href="https://github.com/pySRURGS/pyGOURGS" target ="_blank">https://github.com/pySRURGS/pyGOURGS</a>
**Version:** 0.1
