### 2020-02-05 #2074 @nicoguaro
I tried running the example with

    $ python ant.py

with output

```bash
Traceback (most recent call last):
  File "ant.py", line 118, in <module>
    output_db = args[0]
IndexError: list index out of range
```

with

$ python ant.py output

I get

```bash
Traceback (most recent call last):
  File "ant.py", line 119, in <module>
    n_iters = int(args[1])
IndexError: list index out of range
```

and, finally, with

    $ python ant.py output 10

I didn't have any error message
