### 2020-02-05 #2074 @pySRURGS
The results would then be saved to the file called "output" without any extension.
