### 2020-02-05 #2074 @pySRURGS
There would have been 10 candidate solutions considered in that run.
