### 2020-02-05 #2074 @nicoguaro
Yes, that's what I understood by looking at the source code. But those things should be explained in the documentation.
