### 2020-02-06 #2074 @iljah
@pySRURGS it would be helpful if you'd mention how this relates to e.g. https://www.gnu.org/software/gsl/doc/html/multimin.html. Does this solve a more general problem or is this more efficient than GSL or..?
