### 2020-02-06 #2074 @iljah
As @nicoguaro mentioned, @pySRURGS  please add at least one simple example to documentation. Also it would be helpful for example programs to print usage instructions when used incorrectly, etc.
