### 2020-02-07 63b69c8
adding documentation on artificial ant
