+++ README.md
@@ -43,9 +43,48 @@ pip install -r requirements.txt --user
 
 ### An Example: The Artificial Ant Problem
 
-The artificial ant problem is one in which we identify a search strategy for an ant sea. rching 
-for breadcrumbs to eat. The crumbs are distributed in a path within a 32 x 32 grid. The ant 
-starts at the top left of the grid, and takes steps according to the search strategy. The ant is 
+The artificial ant problem is one in which we identify a search strategy for an ant searching 
+for breadcrumbs to eat. The crumbs are distributed in a path within a 32 x 32 grid. 
+Included in our `/examples/` folder, there are three maps, the `johnmuir_trail.txt`, 
+`losaltoshills_trail.txt`, and the `santafe_trail.txt`. By default, the example in `examples/ant.py` 
+runs against the `johnmuir_trail.txt`. In the `johnmuir` grid 
+shown below, `S` denotes the ant's starting position, `#` denotes a piece of bread, and `.` 
+denotes a space without food.
+
+S##########.....................
+..........#.....................
+..........#.....................
+..........#.....................
+####......#....#................
+...#......#..............#######
+...#......#.............#.......
+...#......#...#.........#.......
+...#......#.#...........#.......
+...#......#.............#.......
+...########.............#.......
+...................#####........
+...........#......#.............
+..............#...#.............
+..................#.............
+...............#..#.............
+............#.....#.............
+..................#.............
+...........#....................
+........#.......................
+..................#.............
+..................#.............
+.......#..........#.............
+.....#............#.............
+..................#.............
+....#.............#.............
+....#...........................
+....#...........................
+....#..####.######..............
+................................
+................................
+................................
+
+The ant takes steps according to the search strategy. The ant is 
 permitted three base operations, 
 
 1. MOVE forward
@@ -60,10 +99,6 @@ permitted. These functions are PROGN2, PROGN3, and IF_FOOD_AHEAD.
 
 Each base operation takes one unit of time to perform. In the included example, the simulation stops running after 
 600 time units.
-
-Included in our `/examples/` folder, there are three maps, the `johnmuir_trail.txt`, 
-`losaltoshills_trail.txt`, and the `santafe_trail.txt`. By default, the example in `examples/ant.py` 
-runs against the `johnmuir_trail.txt`. 
 
 ## API
 
