+++ README.md
@@ -52,7 +52,9 @@ shown below, `S` denotes the ant's starting position, `#` denotes a piece of bre
 denotes a space without food.
 
 S##########.....................
+
 ..........#.....................
+
 ..........#.....................
 ..........#.....................
 ####......#....#................
