### 2020-02-07 a7befd0
draw the plot
