### 2020-02-07 2432e8a
changed output of evalArtificialAnt from tuple to flaot`
