### 2020-02-07 4d1a232
added feature to check for soln in db before running
