### 2020-02-07 e6624e6
defining goodness
