### 2020-02-07 65b8426
developing method to query previous results
