+++ requirements.txt
@@ -4,3 +4,4 @@ pytest==5.1.1
 pytest-cov==2.7.1
 methodtools==0.1.2
 sqlitedict==1.6.0
+tabulate==0.8.3
