### 2020-02-07 125d8b5
added deterministic mode
