+++ examples/ant.py
@@ -120,6 +120,7 @@ if __name__ == "__main__":
     parser.add_argument("-num_trees", help="pyGOURGS iterates through all the possible trees using an enumeration scheme. This argument specifies the number of trees to which we restrict our search.", type=int, default=10000)
     parser.add_argument("-num_iters", help="An integer specifying the number of search strategies to be attempted in this run", type=int, default=1000)
     parser.add_argument("-freq_print", help="An integer specifying how many strategies should be attempted before printing current job status", type=int, default=10)
+    parser.add_argument("-deterministic", help="should algorithm be run in deterministic manner?", type=bool, default=False)
     if len(sys.argv) < 2:
         parser.print_usage()
         sys.exit(1)
@@ -128,11 +129,14 @@ if __name__ == "__main__":
     n_iters = arguments.num_iters
     maximum_tree_complexity_index = arguments.num_trees
     frequency_printing = arguments.freq_print
+    deterministic = arguments.deterministic
     with open("./johnmuir_trail.txt") as trail_file:
         ant.parse_matrix(trail_file)
     max_score = 0
     iter = 0
-    for soln in enum.uniform_random_global_search(maximum_tree_complexity_index, n_iters):
+    for soln in enum.uniform_random_global_search(maximum_tree_complexity_index, 
+                                                  n_iters, 
+                                                  deterministic=deterministic):
         iter = iter + 1 
         score = evalArtificialAnt(soln)
         pg.save_result_to_db(output_db, score, soln)
