### 2020-02-07 078ee0f
added determinsitic and exhaustive modes to the ant example
