### 2020-02-07 71809ee
included example output from deterministic mode
