### 2020-02-07 #2074 @pySRURGS
@iljah @nicoguaro
I have made the corrections you suggested. Please review the readme as it has been edited. Please review the ant.py example file as it has been greatly altered. It now has an interface that you can use from command line and this is discussed in the readme.
