### 2020-02-08 #2074 @nicoguaro
I guess that the line

    $ winpty python ant.py -h

is for Windows (?).

I ran

    $ python ant.py

and obtained

```bash
Traceback (most recent call last):
  File "ant.py", line 101, in <module>
    pset = pg.PrimitiveSet()
AttributeError: module 'pyGOURGS' has no attribute 'PrimitiveSet'
```
