### 2020-02-08 #2074 @pySRURGS
@nicoguaro
Yes winpty is for windows and has no effect on my code's function.
That is odd, and your error message is not that helpful. If the script is in the current directory and you are running it from terminal then I am not sure the cause of the error. I don't want to come across dismissive but did you `git pull` before running?
