### 2020-02-08 #2074 @pySRURGS
@nicoguaro I ran a fresh install of the software on linux and ran
`python ant.py` and did not encounter the issue.
please check that you ran the script from within the /examples/ directory, as I import the pyGOURGS package using a relative path.
