### 2020-02-08 #2074 @pySRURGS
I added commentary to the readme about running ant.py when in the /examples directory because of the relative path import of pyGOURGS
