### 2020-02-09 0898e87
Update README.md
removed sentence
