### 2020-02-11 2015c38
Added JOSS badge
