### 2020-02-12 005846c
latest version of script for fig 4
