+++ experiments/determining_trees_with_n_or_fewer_nodes.py
@@ -27,9 +27,7 @@ trees = []
 for i in range(0,N):
     tree = enum.ith_n_ary_tree(i)
     n_nodes = pg.count_nodes_in_tree(tree)
-    if i % 10000 == 0:
-        print(i, n_nodes)        
-    else:    
+    if i % 100:
         values_of_i.append(i)
         number_of_configs_at_i.append(n_nodes)
         trees.append(tree)
@@ -39,7 +37,10 @@ y = number_of_configs_at_i
 #plt.scatter(x,y)
 x = np.array(x)
 y = np.array(y)
-sns.kdeplot(x, y, cmap="Blues", shade=True, shade_lowest=False)
-plt.ylabel("number of nodes")
-plt.xlabel("pyGOURGS enumeration iterable")
+plt.figure(figsize=(5,3))
+plt.scatter(x, y, color='purple', alpha=0.1)
+sns.kdeplot(x, y, cmap="Blues")
+plt.ylabel("number of nodes in tree")
+plt.xlabel("index of enumeration scheme")
+plt.tight_layout()
 plt.show()
