+++ experiments/determining_trees_with_n_or_fewer_nodes.py
@@ -27,10 +27,9 @@ trees = []
 for i in range(0,N):
     tree = enum.ith_n_ary_tree(i)
     n_nodes = pg.count_nodes_in_tree(tree)
-    if i % 100:
-        values_of_i.append(i)
-        number_of_configs_at_i.append(n_nodes)
-        trees.append(tree)
+    values_of_i.append(i)
+    number_of_configs_at_i.append(n_nodes)
+    trees.append(tree)
 
 x = list(values_of_i)
 y = number_of_configs_at_i
@@ -39,7 +38,6 @@ x = np.array(x)
 y = np.array(y)
 plt.figure(figsize=(5,3))
 plt.scatter(x, y, color='purple', alpha=0.1)
-sns.kdeplot(x, y, cmap="Blues")
 plt.ylabel("number of nodes in tree")
 plt.xlabel("index of enumeration scheme")
 plt.tight_layout()
