### 2020-02-15 e975b31
script to run brute force
