+++ experiments/run_brute_force_first_200_trees_ant.sh
@@ -1,0 +1,3 @@
+cd ..
+cd examples
+python3 ant.py -num_trees 200 -exhaustive True -multiprocessing True ~/brute_force_200.db
