### 2020-02-16 c9e6206
a bit of editing with exhaustive mode
