### 2020-02-16 4317a06
install script for aws
