+++ experiments/install.sh
@@ -1,0 +1,4 @@
+sudo apt update
+sudo apt install python3
+sudo apt install python3-pip
+sudo apt install git
