### 2020-02-17 3b985a8
print enumeration progress
