+++ README.md
@@ -141,8 +141,11 @@ to the help. Make sure to execute `ant.py` only when the current working directo
 is `/examples` because the script imports `pyGOURGS.py` using relative paths.
 
 ```
-$ winpty python ant.py  -h
-usage: ant.py [-h] [-num_trees NUM_TREES] [-num_iters NUM_ITERS]                                                                                                                                                                             [-freq_print FREQ_PRINT] [-deterministic DETERMINISTIC]                                                                                                                                                                        [-exhaustive EXHAUSTIVE] [-multiprocessing MULTIPROCESSING]                                                                                                                                                                    output_db
+$ winpty python ant.py -h
+usage: ant.py [-h] [-num_trees NUM_TREES] [-num_iters NUM_ITERS]
+              [-freq_print FREQ_PRINT] [-deterministic DETERMINISTIC]
+              [-exhaustive EXHAUSTIVE]
+              output_db
 
 positional arguments:
   output_db             An absolute filepath where we save results to a SQLite
