### 2020-02-17 2b4fdd7
updated test to include the uniform random search itself
