+++ test.py
@@ -38,12 +38,12 @@ class TestSymbolicRegression(unittest.TestCase):
 
     def setUp(self):
         self.pset = pg.PrimitiveSet()
-        self.pset.add_operator(add, 2)
-        self.pset.add_operator(sub, 1)
-        self.pset.add_operator(truediv, 3)
-        self.pset.add_operator(mul, 1)
-        self.pset.add_variable(1)
-        self.pset.add_variable(0)
+        self.pset.add_operator('add', 2)
+        self.pset.add_operator('sub', 1)
+        self.pset.add_operator('truediv', 3)
+        self.pset.add_operator('mul', 1)
+        self.pset.add_variable('x')
+        self.pset.add_variable('y')
         self.enum = pg.Enumerator(self.pset)
 
     def test_count_unique_trees(self):
@@ -106,6 +106,18 @@ class TestSymbolicRegression(unittest.TestCase):
         self.assertEqual(self.enum.calculate_S_i(4),2)
         self.assertEqual(self.enum.calculate_S_i(5),4)
 
+    def test_uniform_random_global_search(self):
+        solns = []
+        for soln in self.enum.uniform_random_global_search(10000, 10):
+            solns.append(soln)
+        self.assertEqual(len(solns), len(list(set(solns))), 10)
+        soln = self.enum.uniform_random_global_search_once(10000)
+        self.assertEqual(type(soln), str)
+        solns = []
+        for soln in self.enum.exhaustive_global_search(2,5):
+            solns.append(soln)
+        self.assertEqual(len(solns), 5)
+        
 if __name__ == '__main__':
     unittest.main()
 
