### 2020-02-18 73713f4
fixed db tests and removed commit bug
