### 2020-02-18 c19e7da
removed if__name__ main
