### 2020-02-18 cfa67f2
adding more entries to db
