+++ pyGOURGS/pyGOURGS.py
@@ -867,10 +867,7 @@ def initialize_db(path_to_db):
         for the 'best_result' key
     '''
     with SqliteDict(path_to_db, autocommit=True) as results_dict:
-        try:
-            results_dict['best_result']
-        except KeyError:
-            results_dict['best_result'] = Result(None, None)
+        results_dict['best_result'] = None        
     return
     
 def save_result_to_db(path_to_db, result, input):
