### 2020-02-18 7489f96
fixing tst issue init db
