### 2020-02-18 3d2294b
made sure default printout was only 1
