### 2020-02-18 cdb13f8
removed print_grdi
