### 2020-02-19 ebd5a35
hyperoptimized performance of ant.py multiprocessing
