### 2020-02-19 8755eda
added some lines to install script for ease
