+++ experiments/install.sh
@@ -2,3 +2,6 @@ sudo apt update
 sudo apt install python3
 sudo apt install python3-pip
 sudo apt install git
+git clone https://github.com/pySRURGS/pyGOURGS.git
+cd pyGOURGS 
+pip3 install -r requirements.txt --user 
