### 2020-03-04 #2074 @nicoguaro
@arokem I have completed my review and I would suggest publication. I do have a suggestion to the documentation, though:

> Add tutorials/examples to it. Currently it only provides an API documentation without explanations. I would suggest to start with the ant example and expand it with explanations.

@pySRURGS, I also have some comments regarding the paper. You can find an annotated PDF [here](https://www.dropbox.com/s/flshpkopb3brckv/10.21105.joss.02074.pdf?dl=0).
