### 2020-03-04 #2074 @sohrabtowfighi
> @arokem I have completed my review and I would suggest publication. I do have a suggestion to the documentation, though:
>
> > Add tutorials/examples to it. Currently it only provides an API documentation without explanations. I would suggest to start with the ant example and expand it with explanations.

Could you help me by clarifying this? The API documentation is provided with comments for just about every function and class. The artificial ant example is explored in detail in the readme.md and a basic example run is shown. Do you think I should add more examples on how to use the bash interface?
