### 2020-03-04 #2074 @nicoguaro
Yes, I think so.

The API documentation and examples/tutorials serve to different types of users. For newcomers to your software, the API documentation will be less useful than step-by-step explanations. The opposite happens for competent users.
