### 2020-03-06 89f202a
added bash usage example
