### 2020-03-06 98d4141
added example usage of bash interface
