### 2020-03-06 fda98d4
some test depends on default value of print for resultlist to be small
