### 2020-03-06 #2074 @sohrabtowfighi
Ok @nicoguaro I think it is ready for re-review. I have incorporated most your edits into the paper. I also added several examples and subheaders in the description of the artificial ant example in the readme.md
