### 2020-03-13 #2074 @arokem
@nicoguaro : I see that "example usage" is still unchecked.  Just making sure: is this ready for publication from your point of view?
