### 2020-03-13 #2074 @arokem
I see now. The DOI suggested by whedon is incorrect (points to another resource). I think that you can leave this as is.

In the meanwhile, while we are waiting for @nicoguaro's final approval, could you please go ahead and proceed with the following steps:

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g. figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata, this includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it); you may also add the authors' ORCID.

Once we have these items, and @nicoguaro's confirmation, we can move ahead with the acceptance of the paper.
