### 2020-03-13 #2074 @nicoguaro
> @nicoguaro : I see that "example usage" is still unchecked. Just making sure: is this ready for publication from your point of view?

Yes, it is checked now.
