+++ papers/JOSS/paper.bib
@@ -1,7 +1,7 @@
 @article{wolpert1997no,
   title={No free lunch theorems for optimization},
   author={Wolpert, David H and Macready, William G and others},
-  journal={IEEE transactions on evolutionary computation},
+  journal={IEEE Transactions on Evolutionary Computation},
   volume={1},
   number={1},
   pages={67--82},
@@ -62,7 +62,7 @@ doi = {10.1186/s13040-018-0164-x}
 @InProceedings{langdon:1998:antspace,
 author = "W. B. Langdon and R. Poli",
 title = "Why Ants are Hard",
-booktitle = "Genetic Programming 1998: Proceedings of the Third Annual Conference",
+booktitle = {{Proceedings of the Third Annual Conference on Genetic Programming (GP98)}},
 year = "1998",
 editor = "John R. Koza and Wolfgang Banzhaf and Kumar Chellapilla and Kalyanmoy Deb and Marco Dorigo and David B. Fogel and Max H. Garzon and David E. Goldberg and Hitoshi Iba and Rick Riolo",
 pages = "193--201",
@@ -73,10 +73,12 @@ URL = "http://www.cs.ucl.ac.uk/staff/W.Langdon/ftp/papers/WBL.antspace\_gp98.pdf
 }
 
 @article{towfighi2019pysrurgs,
-  title={pySRURGS-a python package for symbolic regression by uniform random global search},
+  title={pySRURGS - a python package for symbolic regression by uniform random global search},
   author={Towfighi, Sohrab},
   journal={The Journal of Open Source Software},
   volume={4},
   year={2019},
+  number = {41},
+  pages = {1675},
   doi = {10.21105/joss.01675}
 }
