### 2020-03-19 #2074 @danielskatz
👋 @pySRURGS - please merge https://github.com/pySRURGS/pyGOURGS/pull/2 and https://github.com/pySRURGS/pyGOURGS/pull/3 or let me know what you disagree with. These are changes based on proof-reading the final draft paper.
