### 2020-03-20 #2074 @pySRURGS
@danielskatz , I have merged all suggested changes.
